// src/pages/modules/PMB/entrance/content/data/types.ts
export interface IContentEntrance {
  id_jalur_masuk_konten: string
  id_satuan_organisasi: string
  id_jalur_masuk: string
  judul_konten: string
  slug: string
  isi_konten: string
  urutan: number
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
  nama_satuan_organisasi: string
  nama_jalur_masuk: string
  nama_user_created: string
  nama_user_updated: string
}